requirements.txt
streamlit
pandas
requests
pillow
google-cloud-vision
